backend/django_admin/requirements.txt
Django==4.1.7
django-celery-beat==2.4.0
redis==4.5.1
python-dotenv==1.0.0
django-split-settings==1.2.0
gunicorn==20.1.0
psycopg2==2.9.3
requests==2.28.2
djangorestframework==3.14.0